requirements.txt
pandas
numpy
openai
jupyter
ipykernel
FlagEmbedding
langchain
langchain-community
unstructured
Markdown
tqdm
faiss-gpu
ollama
fastapi
uvicorn
marko